xai/requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
groq>=0.4.1
openai>=1.0.0
requests>=2.0.0

# Optionnel (pour connexion DB Supabase)
# sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.9
# supabase>=2.3.0